refactor(sections): Sections 도메인으로 로직 옮기기
