+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -15,9 +15,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    // TODO Sections 일급 컬렉션 생성
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = Sections.empty();
 
     protected Line() {}
 
@@ -29,8 +28,7 @@ public class Line extends BaseEntity {
     private Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        // TODO addSection 호출
-        this.sections.add(Section.of(this, upStation, downStation, distance));
+        this.addSection(Section.of(this, upStation, downStation, distance));
     }
 
     public static Line empty() {
@@ -51,112 +49,16 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(Section section) {
-        // TODO Sections 에서 처리------------------------
-        List<Station> stations = getStations();
-        boolean isUpStationExisted = stations.stream().anyMatch(it -> it == section.getUpStation());
-        boolean isDownStationExisted = stations.stream().anyMatch(it -> it == section.getDownStation());
-
-        if (isUpStationExisted && isDownStationExisted) {
-            throw new RuntimeException("이미 등록된 구간 입니다.");
-        }
-
-        if (!stations.isEmpty() && stations.stream().noneMatch(it -> it == section.getUpStation()) &&
-                stations.stream().noneMatch(it -> it == section.getDownStation())) {
-            throw new RuntimeException("등록할 수 없는 구간 입니다.");
-        }
-
-        if (stations.isEmpty()) {
-            getSections().add(section);
-            return;
-        }
-
-        if (isUpStationExisted) {
-            this.sections.stream()
-                    .filter(it -> it.getUpStation() == section.getUpStation())
-                    .findFirst()
-                    .ifPresent(it -> it.updateUpStation(section.getDownStation(), section.getDistance()));
-
-            this.sections.add(section);
-        } else if (isDownStationExisted) {
-            this.sections.stream()
-                    .filter(it -> it.getDownStation() == section.getDownStation())
-                    .findFirst()
-                    .ifPresent(it -> it.updateDownStation(section.getUpStation(), section.getDistance()));
-
-            this.sections.add(section);
-        } else {
-            throw new RuntimeException();
-        }
-        // TODO Sections 에서 처리------------------------
-    }
-
-    public List<Station> getStations() {
-        // TODO Sections 에서 처리------------------------
-        if (this.sections.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = this.sections.stream()
-                    .filter(it -> it.getUpStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
-        // TODO Sections 에서 처리------------------------
-    }
-
-    private Station findUpStation() {
-        // TODO Sections 에서 처리------------------------
-        Station downStation = this.sections.get(0).getUpStation();
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = this.sections.stream()
-                    .filter(it -> it.getDownStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
-        // TODO Sections 에서 처리------------------------
+        this.sections.add(section);
+        section.assignLine(this);
     }
 
     public void removeStation(Station station) {
-        // TODO Sections 에서 처리------------------------
-        if (this.sections.size() <= 1) {
-            throw new RuntimeException();
-        }
+        this.sections.removeStation(station, this);
+    }
 
-        Optional<Section> upLineStation = this.sections.stream()
-                .filter(it -> it.getUpStation() == station)
-                .findFirst();
-        Optional<Section> downLineStation = this.sections.stream()
-                .filter(it -> it.getDownStation() == station)
-                .findFirst();
-
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            Station newUpStation = downLineStation.get().getUpStation();
-            Station newDownStation = upLineStation.get().getDownStation();
-            int newDistance = upLineStation.get().getDistance() + downLineStation.get().getDistance();
-            this.sections.add(Section.of(this, newUpStation, newDownStation, newDistance));
-        }
-
-        upLineStation.ifPresent(it -> this.sections.remove(it));
-        downLineStation.ifPresent(it -> this.sections.remove(it));
-        // TODO Sections 에서 처리------------------------
+    public List<Station> getStations() {
+        return this.sections.getStations();
     }
 
     public Long getId() {
@@ -172,7 +74,7 @@ public class Line extends BaseEntity {
     }
 
     public List<Section> getSections() {
-        return this.sections;
+        return this.sections.get();
     }
 
     @Override
